phase9.
Replace Buttreknife into Databinding, because Butterknife was not
as usable I though and attract to OBSERVABLE of Databinding.

Tried to build project successfully, but still views were not
updated according to commands.
(I think it is because of first Query command is not worked well at
the point of onCreate..)

will focus on TODO comments.